All, NewBoard.java -> Board.java
Renamed the NewBoard.java to make the name more sensible. Some lines in
StartGame.java needed to be changed: In the previous version, the board
was canvas based and needed to be attached to a Frame. Now that is not
needed anymore. The lines where the board is attached are removed.