GH & LAK, corrected number display in ResourceNumbers.java
There was a 6 missing in the top left. It got overwritten by the ''desert'' tile which was in the wrong position.
As the desert tile is not needed, we decided to // comment the lines and leave it as is for now.
On the 5th row, the 10 and 11 numbers were conflicted. We fixed the problem and now all numbers display properly on the board.